add more sample
